Jobs/requirement.txt
pip install \
  requests \
  beautifulsoup4 \
  psycopg2-binary \
  python-dotenv \
  google-api-python-client \
  google-auth \
  google-auth-oauthlib \
  google-auth-httplib2 \
  urllib3
